Integrated costs with DDP solver
- Added test_costs.py to test the cost function
- Added test_dubin_dynamics.py to test the analytical jacobians of the dynamics
- Added test_dubin_w_ddp.py as a demo putting everything together
- Fixed several bugs in DDPSolver -- should be good to go!

(Note, added addcopyfighandler dependency, its nice to be able to copy matplotlib figures).

+++ solvers/ocp_interface.py
@@ -10,4 +10,4 @@ class OCP:
     stage_cost: BaseCost
     terminal_cost: BaseCost
     horizon: int
-    dt: int
+    dt: float
